+++ public/cache-items.js
@@ -1,0 +1,35 @@
+'use strict';
+
+(function () {
+  window.cacheItems = (cache) => {
+    return cache.addAll([
+      `./`,
+      `./index.html`,
+      `./bundle.js`,
+      `./css/style.css`,
+      `./fonts/Montserrat-Bold.woff2`,
+      `./fonts/Montserrat-ExtraBold.woff2`,
+      `./fonts/Montserrat-Medium.woff2`,
+      `./fonts/Montserrat-Regular.woff2`,
+      `./fonts/Montserrat-SemiBold.woff2`,
+      `./img/icons/bus.png`,
+      `./img/icons/check-in.png`,
+      `./img/icons/drive.png`,
+      `./img/icons/flight.png`,
+      `./img/icons/restaurant.png`,
+      `./img/icons/ship.png`,
+      `./img/icons/sightseeing.png`,
+      `./img/icons/taxi.png`,
+      `./img/icons/train.png`,
+      `./img/icons/transport.png`,
+      `./img/photos/1.jpg`,
+      `./img/photos/2.jpg`,
+      `./img/photos/3.jpg`,
+      `./img/photos/4.jpg`,
+      `./img/photos/5.jpg`,
+      `./img/header-bg.png`,
+      `./img/header-bg@2x.png`,
+      `./img/logo.png`
+    ]);
+    };
+})();
