+++ public/css/style.css
@@ -1006,3 +1006,30 @@ main {
 .shake {
   animation: shake 0.6s;
 }
+
+.toast-container {
+  position: absolute;
+  z-index: 1000;
+  top: 0;
+  right: 0;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding-top: 0.4em;
+  padding-left: 0.4em;
+  width: 100%;
+  height: 0;
+  font-family: sans-serif;
+  font-size: 16px;
+  line-height: 1.5;
+  box-sizing: border-box;
+}
+
+.toast-item {
+  display: inline-flex;
+  margin-bottom: 0.4em;
+  padding: 0.4em;
+  border-radius: 0.2em;
+  background-color: #fee;
+  color: #900;
+}
