+++ public/index.html
@@ -7,7 +7,6 @@
     <title>Big Trip</title>
 
     <link rel="stylesheet" href="./css/style.css">
-    <link rel="stylesheet" href="../src/utils/toast/toast.css">
   </head>
 
   <body class="page-body">
@@ -34,6 +33,7 @@
       <div class="page-body__container">
       </div>
     </main>
+    <script src="./cache-items.js"></script>
     <script src="./sw.js"></script>
     <script src="./bundle.js"></script>
   </body>
