+++ public/sw.js
@@ -1,6 +1,6 @@
 const CACHE_PREFIX = `big-trip-cache`;
-const CACHE_VER = `v13`;
-const CACHE_NAME = `${CACHE_PREFIX}-${CACHE_VER}`;
+const CACHE_VERSION = `v13`;
+const CACHE_NAME = `${CACHE_PREFIX}-${CACHE_VERSION}`;
 
 const HTTP_STATUS_OK = 200;
 const RESPONSE_SAFE_TYPE = `basic`;
@@ -8,38 +8,7 @@ const RESPONSE_SAFE_TYPE = `basic`;
 self.addEventListener(`install`, (evt) => {
   evt.waitUntil(
       caches.open(CACHE_NAME)
-      .then((cache) => {
-        return cache.addAll([
-          `./`,
-          `./index.html`,
-          `./bundle.js`,
-          `./css/style.css`,
-          `./css/toast.css`,
-          `./fonts/Montserrat-Bold.woff2`,
-          `./fonts/Montserrat-ExtraBold.woff2`,
-          `./fonts/Montserrat-Medium.woff2`,
-          `./fonts/Montserrat-Regular.woff2`,
-          `./fonts/Montserrat-SemiBold.woff2`,
-          `./img/icons/bus.png`,
-          `./img/icons/check-in.png`,
-          `./img/icons/drive.png`,
-          `./img/icons/flight.png`,
-          `./img/icons/restaurant.png`,
-          `./img/icons/ship.png`,
-          `./img/icons/sightseeing.png`,
-          `./img/icons/taxi.png`,
-          `./img/icons/train.png`,
-          `./img/icons/transport.png`,
-          `./img/photos/1.jpg`,
-          `./img/photos/2.jpg`,
-          `./img/photos/3.jpg`,
-          `./img/photos/4.jpg`,
-          `./img/photos/5.jpg`,
-          `./img/header-bg.png`,
-          `./img/header-bg@2x.png`,
-          `./img/logo.png`
-        ]);
-      })
+      .then(wnindow.cacheItems(cache))
   );
 });
